requirements.txt
matplotlib~=3.4.3
numpy~=1.21.2
scikit_learn==0.24.2
SimpleITK==2.1.0
scikit-image~=0.18.3
Pillow~=8.4.0
tensorflow~=2.7.0
pandas~=1.3.4
scikit-learn~=0.24.2
keras~=2.7.0
natsort~=8.0.1